removed electronjs code
