+++ app.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const server = require('./server.js'),
+    http = require('http'),
+    fs = require('fs');
+
+server.startServer(http, fs);
